Create Day 1 #100daysofcode
At times we want the output of a program to be printed in a given specific format. In java we can achieve this by using the printf( ) Method. An important point to note is that System.out.print( ) and System.out.println( ) take a single argument, but the printf( ) method can accept multiple arguments. Java's System.out.printf function can be used to print formatted output.

Here I have solved a question of Hacker Rank using printf() method.